requirements.txt
# Web Framework
flask==3.0.0
flask-sqlalchemy==3.1.1

# LINE Bot SDK
line-bot-sdk==3.6.0

# Type hints support  
typing-extensions>=4.0.0

# AI/ML
google-generativeai>=0.8.0
python-gemini-api>=0.1.0
pillow>=10.0.0
anthropic>=0.25.0
magic_hour>=0.42.0

# HTTP & API
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
jinja2>=3.1.2
tenacity>=8.2.3
ddgs>=9.6.0

# Google APIs (検索機能用) - バージョン統一
google-api-python-client==2.108.0

# Configuration
python-dotenv==1.0.0
watchdog>=2.1.9
PyYAML>=6.0.0

# Database drivers
psycopg2-binary>=2.9.9

# 開発環境専用
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# feedparser==6.0.10
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
# lxml==4.9.3
# RAG用 (LangChain不使用の軽量版)
janome>=0.4.2
scikit-learn>=1.3.0
pandas>=2.0.0

# バッチ処理とOAuth用の追加依存関係
apscheduler==3.10.4
requests-oauthlib==1.3.1
selenium>=4.15.0
webdriver-manager>=4.0.0
schedule>=1.2.0
xmltodict==0.13.0

# Google Photos API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Removed LangChain dependencies due to installation issues
# langchain>=0.2.0
# langchain-core>=0.2.0
# langchain-google-genai>=0.1.0
# langchain-anthropic>=0.1.0
# langchain-openai>=0.1.0
pydantic>=1.10,<3
# psycopg2-binary>=2.9.9  # PostgreSQL not needed without LangChain
yt-dlp
